services/realtime/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
websockets==13.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
psycopg==3.2.9
PyJWT==2.10.1
redis==5.2.1
